+++ addOns/client/src/main/java/org/zaproxy/addon/client/spider/actions/ClickElement.java
@@ -29,6 +29,7 @@ import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.zaproxy.addon.client.spider.ActionWaitStrategy;
 import org.zaproxy.addon.commonlib.ValueProvider;
 import org.zaproxy.zap.utils.Stats;
 
@@ -55,7 +56,8 @@ public class ClickElement extends BaseElementAction {
     }
 
     @Override
-    public boolean run(WebDriver wd, WebElement element, String statsPrefix) {
+    public boolean run(
+            ActionWaitStrategy waitStrategy, WebDriver wd, WebElement element, String statsPrefix) {
         if (!passive) {
             fillInputs(wd.findElements(By.xpath("//input")), getUri().toString(), statsPrefix);
         }